Test/controller

+++ src/main/java/com/tinyhuman/tinyhumanapi/album/controller/dto/AlbumCreate.java
@@ -13,9 +13,9 @@ public record AlbumCreate(
         @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
         LocalDateTime originalCreatedAt,
 
-        Long gpsLat,
+        Double gpsLat,
 
-        Long gptLon
+        Double gptLon
 
         ) {
